requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cohere==4.45
contourpy==1.2.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
fastapi==0.109.2
fastavro==1.9.3
fiona==1.9.5
fonttools==4.48.1
frozenlist==1.4.1
geopandas==0.14.3
google-auth==2.27.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
llmx==0.0.21a0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-venn==0.11.10
mizani==0.9.3
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
openai==1.11.1
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pillow==10.2.0
plotly==5.18.0
plotnine==0.12.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.2
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
statsmodels==0.14.1
tenacity==8.2.3
tiktoken==0.5.2
toolz==0.12.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
wordcloud==1.9.3
yarl==1.9.4
zipp==3.17.0
